requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==23.1.0
boto3==1.26.140
botocore==1.29.140
certifi==2023.5.7
crispy-bootstrap5==0.7
Django==4.1.6
django-crispy-forms==2.0
django-storages==1.13.2
exceptiongroup==1.1.1
factory-boy==3.2.1
Faker==18.7.0
gunicorn==20.1.0
h11==0.14.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
jmespath==1.0.1
outcome==1.2.0
packaging==23.1
Pillow==9.4.0
pluggy==1.0.0
PySocks==1.7.1
pytest==7.3.1
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
s3transfer==0.6.1
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
urllib3==1.26.16
wsproto==1.2.0
